% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_registros.R
\name{get_num_registros}
\alias{get_num_registros}
\title{GET Registros SB}
\usage{
get_num_registros(
  api_key,
  end_point,
  periodo_inicial,
  periodo_final = NULL,
  entidad = NULL,
  tipo_entidad = NULL,
  registros = NULL,
  paginas = 100
)
}
\arguments{
\item{api_key}{Primary Key de la suscripcion.}

\item{end_point}{Endpoint a consultar. Usa 'end_points()' para ver
todas las opciones.}

\item{periodo_inicial}{Fecha en donde incian las series solicitadas}

\item{periodo_final}{Fecha en donde terminan las series. No obligatoria.}

\item{entidad}{ID de la entidad a consultar. No se puede introducir al mismo
tiempo que 'tipo_entidad'. Usa 'entidades()' para ver todos los IDs.}

\item{tipo_entidad}{ID del tipo de entidad a consulta. No se puede introducir
al mimo tiempo que 'entidad'. Usa 'tipo_entidad()' para ver todas
las opciones.}

\item{paginas}{cantidad de paginas a consultar. El numero de registros
depende le las paginas consultadas. 100 páginas por defecto.}
}
\value{
string con el número de registros totales. Esta cantidad de
registros está basada en que la consulta está dividida en 100 páginas.
}
\description{
Permite consultar el numero de registros de una consulta en especifica. Util
para crear cualquier proceso automático.
}
