% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb_get.R
\name{sb_get}
\alias{sb_get}
\title{GET Endpoints SB}
\usage{
sb_get(
  api_key,
  end_point,
  periodo_inicial,
  periodo_final = NULL,
  entidad = NULL,
  tipo_entidad = NULL
)
}
\arguments{
\item{api_key}{Primary Key de la suscripcion.}

\item{end_point}{Endpoint a consultar. Usa 'end_points()' para ver
todas las opciones}

\item{periodo_inicial}{Fecha en donde incian las series solicitadas}

\item{periodo_final}{Fecha en donde terminan las series. No obligatoria.}

\item{entidad}{ID de la entidad a consultar. No se puede introducir al mismo
tiempo que 'tipo_entidad'. Usa 'entidades()' para ver todos los IDs.}

\item{tipo_entidad}{ID del tipo de entidad a consulta. No se puede introducir
al mimo tiempo que 'entidad'. Usa 'tipo_entidad()' para ver todas
las opciones}
}
\value{
Dataframe con los datos del endpoint consultado
}
\description{
Esta funcion carga los datos del endpoint solicitado en formato data frame
}
