% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb_get_v2.R
\name{sb_get_v2}
\alias{sb_get_v2}
\title{GET Endpoints SB}
\usage{
sb_get_v2(
  api_key,
  end_point,
  periodo_inicial,
  periodo_final = NULL,
  entidad = NULL,
  tipo_entidad = NULL,
  paginas = 1,
  registros = 100
)
}
\arguments{
\item{api_key}{Primary Key de la suscripcion.}

\item{end_point}{Endpoint a consultar. Usa 'end_points()' para ver
todas las opciones}

\item{periodo_inicial}{Fecha en donde incian las series solicitadas}

\item{periodo_final}{Fecha en donde terminan las series. No obligatoria.}

\item{entidad}{ID de la entidad a consultar. No se puede introducir al mismo
tiempo que 'tipo_entidad'. Usa 'entidades()' para ver todos los IDs.}

\item{tipo_entidad}{ID del tipo de entidad a consulta. No se puede introducir
al mimo tiempo que 'entidad'. Usa 'tipo_entidad()' para ver todas
las opciones}

\item{{paginas}}{{Numero de paginas a consultar. La cantidad de paginas
dependera del numero de registros que se consulten.Su valor por defecto es de
1}}

\item{{registros}}{{Numero de registros a consultar. La cantidad de registros
dependera del numero de pagina que se consulten. Su valor por defecto es de
100}}
}
\value{
Dataframe con los datos del endpoint consultado
}
\description{
Esta nueva version permite utilizar la V2 del API. Ahora utiliza paginacion para
hacer las consultas indicando la cantidad de paginas y de registros a
consultar
}
