% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipo_entidades.R
\name{tipo_entidades}
\alias{tipo_entidades}
\title{Muestra los codigos de los tipos de entidades para hacer las consultas}
\usage{
tipo_entidades()
}
\description{
Muestra los codigos de los tipos de entidades para hacer las consultas
}
